+++ src/main/java/com/gwangya/user/service/UserService.java
@@ -4,6 +4,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.gwangya.global.exception.NoExistEntityException;
 import com.gwangya.global.util.ConvertUtil;
 import com.gwangya.user.domain.User;
 import com.gwangya.user.dto.UserCreateCommand;
@@ -12,8 +13,8 @@ import com.gwangya.user.repository.UserRepository;
 
 import lombok.RequiredArgsConstructor;
 
-@Service
 @RequiredArgsConstructor
+@Service
 public class UserService {
 
 	private final PasswordEncoder passwordEncoder;
@@ -24,5 +25,12 @@ public class UserService {
 	public UserDto createUser(final UserCreateCommand userCreateCommand) {
 		User savedUser = userRepository.save(User.of(userCreateCommand, passwordEncoder, userRepository));
 		return ConvertUtil.convert(savedUser, UserDto.class);
+	}
+
+	@Transactional(readOnly = true)
+	public UserDto searchUserById(final Long userId) {
+		User user = userRepository.findById(userId)
+			.orElseThrow(() -> new NoExistEntityException("존재하지 않는 유저입니다."));
+		return ConvertUtil.convert(user, UserDto.class);
 	}
 }
